优先判断扣减 再发放物品代码完成

+++ Basic/src/main/java/com/game/template/basic/id/enums/GlobalIdType.java
@@ -21,7 +21,7 @@ public enum  GlobalIdType {
 		return type;
 	}
 
-	public int generatorId(){
+	public long generatorId(){
 		return GlobalIdBuilderService.instance.getGlobalIdBuilderBo(getType());
 	}
